Split collection service unit tests into separate files
